1. change Makefile in ctest/soc(add start.S for sp init)

+++ src/main/resources/ctest/soc/device/sdt.c
@@ -1,6 +1,6 @@
 #include "sdt.h"
 
-void SDT_Write(SDT_TypeDef* SDTx, uint16_t PortVal)
+void SDT_Write(SDT_TypeDef* SDTx, uint32_t PortVal)
 {
   /* Check the parameters */
   assert_param(IS_SDT_ALL_PERIPH(SDTx));
